fix: address documentation and naming comments
